add support for adding/removing CardFragment
