+++ app/src/main/java/de/fraunhofer/fokus/OpenMobileNetworkToolkit/Iperf3/Fragments/Iperf3Fragment.java
@@ -1,0 +1,65 @@
+/*
+ *  SPDX-FileCopyrightText: 2023 Peter Hasse <peter.hasse@fokus.fraunhofer.de>
+ *  SPDX-FileCopyrightText: 2023 Johann Hackler <johann.hackler@fokus.fraunhofer.de>
+ *  SPDX-FileCopyrightText: 2023 Fraunhofer FOKUS
+ *
+ *  SPDX-License-Identifier: BSD-3-Clause-Clear
+ */
+
+package de.fraunhofer.fokus.OpenMobileNetworkToolkit.Iperf3.Fragments;
+
+import android.content.Context;
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.Button;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.viewpager2.widget.ViewPager2;
+
+import de.fraunhofer.fokus.OpenMobileNetworkToolkit.Iperf3.Fragments.Input.Iperf3CardAdapter;
+import de.fraunhofer.fokus.OpenMobileNetworkToolkit.R;
+import de.fraunhofer.fokus.OpenMobileNetworkToolkit.Preferences.SharedPreferencesGrouper;
+
+public class Iperf3Fragment extends Fragment {
+    private static final String TAG = "iperf3InputFragment";
+
+    private View v;
+    private SharedPreferencesGrouper spg;
+    private Button sendBtn;
+    private Button instancesBtn;
+    private Context ct;
+    private ViewPager2 viewPager;
+
+    @Override
+    public void onPause() {
+        super.onPause();
+    }
+
+    @Override
+    public void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup parent, Bundle savedInstanceState) {
+        v = inflater.inflate(R.layout.fragment_iperf3_input, parent, false);
+        this.viewPager = v.findViewById(R.id.iperf3_viewpager);
+        int numPages = 3; // Set this to the number of pages you want
+        Iperf3CardAdapter adapter = new Iperf3CardAdapter(getActivity());
+        this.viewPager.setAdapter(adapter);
+        this.ct = requireContext();
+        this.spg = SharedPreferencesGrouper.getInstance(this.ct);
+        this.sendBtn = v.findViewById(R.id.iperf3_send);
+
+        return v;
+    }
+
+    @Override
+    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
+        super.onViewCreated(view, savedInstanceState);
+    }
+}
