+++ app/src/main/java/de/fraunhofer/fokus/OpenMobileNetworkToolkit/Iperf3/Iperf3ResultsDataBase.java
@@ -21,7 +21,7 @@ import androidx.room.RoomDatabase;
 public abstract class Iperf3ResultsDataBase extends RoomDatabase {
     private static volatile Iperf3ResultsDataBase INSTANCE;
 
-    static Iperf3ResultsDataBase getDatabase(final Context context) {
+    public static Iperf3ResultsDataBase getDatabase(final Context context) {
         if (INSTANCE == null) {
             synchronized (Iperf3ResultsDataBase.class) {
                 if (INSTANCE == null) {
