+++ app/src/main/res/layout/fragment_iperf3_card.xml
@@ -7,6 +7,7 @@
     android:orientation="vertical"
     android:padding="16dp">
     <LinearLayout
+        android:id="@+id/iperf3_header"
         android:layout_height="wrap_content"
         android:layout_width="match_parent"
         android:orientation="horizontal">
@@ -23,7 +24,7 @@
             android:textSize="18sp"/>
 
         <Button
-            android:id="@+id/close_button"
+            android:id="@+id/iperf3_close_button"
             android:layout_width="50dp"
             android:layout_height="wrap_content"
             android:layout_gravity="center_vertical"
