+++ app/src/main/res/layout/fragment_iperf3_list.xml
@@ -15,7 +15,7 @@
     android:exported="true"
     android:background="?android:attr/selectableItemBackground"
     android:id="@+id/iperf3_list_fragment"
-    tools:context=".Iperf3.Iperf3ListFragment" >
+    tools:context=".Iperf3.Fragments.Output.Iperf3ListFragment" >
         <com.google.android.material.floatingactionbutton.FloatingActionButton
             android:id="@+id/iperf3_upload_button"
             android:layout_width="106dp"
